// src/app/models/timesheet.model.ts
// class with properties

export class Timesheetitems {
    department: number = 2;
    wp: string = "";
    total: number = 0;
    monday: number = 0;
    tuesday: number = 0;
    wednesday: number = 0;
    thursday: number = 0;
    friday: number = 0;
    saturday: number = 0;
    sunday: number = 0;
    notes: string = "";

}